refactor: consolidate API clients into one canonical client
Merge sdk.ts, api-client.ts and redeem.ts into create-bridgelet-client.ts,
trim bridgelet.ts to types only, and document the client in the README.
No route or behavior changes.

Closes #228
