+++ frontend/README.md
@@ -50,3 +50,11 @@ Reference Next.js UI for initiating and claiming crypto payments.
   ```bash
   npm run build
   ```
+
+## API client
+
+All calls to the `bridgelet-sdk` backend go through a single canonical client:
+**`lib/create-bridgelet-client.ts`**. Import the `BridgeletClient` class (or the
+`getClaimDetails` / `createPaymentIntent` / `redeemClaim` helpers) from there — do not
+add new API-calling modules or reintroduce parallel clients. Request/response types
+live in `lib/bridgelet.ts` (auto-generated via `npm run generate:types`).
