+++ frontend/components/claim-status-card.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from 'react';
 import { RateLimitBanner } from '@/components/rate-limit-banner';
-import { RateLimitError } from '@/lib/redeem';
+import { RateLimitError } from '@/lib/create-bridgelet-client';
 import { ChainSelector } from '@/components/chain-selector';
