+++ frontend/lib/bridgelet.ts
@@ -74,38 +74,3 @@ export interface ApiError {
   statusCode: number;
 }
 
-import {
-  BridgeletClient,
-  type BridgeletClientOptions,
-} from '@/lib/create-bridgelet-client';
-
-export {
-  BridgeletClient,
-  type BridgeletClientOptions,
-};
-
-let _defaultClient: BridgeletClient | null = null;
-
-function defaultClient(): BridgeletClient {
-  if (!_defaultClient) {
-    _defaultClient = new BridgeletClient();
-  }
-  return _defaultClient;
-}
-
-export function getClaimDetails(token: string): Promise<ClaimDetailsResponse> {
-  return defaultClient().getClaimDetails(token);
-}
-
-export function createPaymentIntent(
-  data: CreatePaymentIntentRequest,
-): Promise<CreatePaymentIntentResponse> {
-  return defaultClient().createPaymentIntent(data);
-}
-
-export function redeemClaim(
-  token: string,
-  data: RedeemClaimRequest,
-): Promise<RedeemClaimResponse> {
-  return defaultClient().redeemClaim(token, data);
-}
